requirements.txt
streamlit
tensorflow-cpu
numpy
pandas
Pillow
fpdf